Added flatfielding_stitcher and cleaned up some

+++ improc/improc/flatfield.py
@@ -1,11 +1,15 @@
+from typing import Iterable
 import numpy as np
 import pathlib
+import itertools
+
+from common import WellSpec
 
 from PIL import Image
 
 def background_from_paths(paths: list[pathlib.Path]) -> np.ndarray:
     images = np.array([np.array(Image.open(path)) for path in paths])
-    return np.median(images, axis=0, overwrite_input=True)
+    return np.median(images, axis=0, overwrite_input=True).astype(images.dtype)
 
 def trunc_sub(a1: np.ndarray, a2: np.ndarray) -> np.ndarray:
     """
